iam-demo/flask-app/requirements.txt
Flask
python-jose[cryptography]
requests